requirements.txt
Flask==0.12.2
Flask-GraphQL==2.0.0
graphene-mongo
mongomock==3.14.0
mongoengine
python-dotenv==0.10.2
